reimplemented scripts

+++ Stratega/include/Stratega/Agent/ActionScripts/AttackClosestOpponentScript.h
@@ -8,7 +8,6 @@
 namespace SGA {
 	class AttackClosestOpponentScript : public BaseActionScript
 	{
-
 	public:
 		AttackClosestOpponentScript() : BaseActionScript(){};
 
@@ -17,7 +16,7 @@ namespace SGA {
 		[[nodiscard]] std::string toString() const override { return "AttackClosestOpponentScript";  };
 
 	private:
-		static int minimalDistanceToOpponents(const Vector2i position, std::map<int, Vector2i>& unitPositions, const std::set<int>& opponentUnits);
+		static double minimalDistanceToOpponents(const Vector2f position, std::map<int, Vector2f>& unitPositions, const std::set<int>& opponentUnits);
 
 	};
